move function and test to utils
